Added test for handling 64 bit integers as the ability to add higher numbers allows the possibility of growing past max 32 bit int.  Processing cost is minimal.

+++ VasysRomanNumeralsKata/FromRomanNumeral.cs
@@ -6,7 +6,7 @@ namespace VasysRomanNumeralsKata
 {
     public class FromRomanNumeral
     {
-        public int GenerateArabicRepresentation(string romanNumeral)
+        public long GenerateArabicRepresentation(string romanNumeral)
         {
             return 10;
         }
